Some improvements for rate limiter configuration

+++ config/api.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 return [
     'rate_limiter' => [
         'enabled' => env('API_RATE_LIMITER_ENABLED', true),
-        'attempts' => env('API_RATE_LIMITER_ATTEMPTS_PER_MIN', 60),
-        'expires' => env('API_RATE_LIMITER_EXPIRES_IN_MIN', 1),
+        'attempts' => (int) env('API_RATE_LIMITER_ATTEMPTS_PER_MIN', 60),
+        'expires' => (int) env('API_RATE_LIMITER_EXPIRES_IN_MIN', 1),
     ],
 ];
